Introducing the HUD layer rendering pass into it s own task and file
