+++ libraries/render-utils/src/RenderCommonTask.h
@@ -77,16 +77,6 @@ protected:
     bool _opaquePass { true };
 };
 
-class CompositeHUD {
-public:
-    // IF specified the input Framebuffer is actively set by the batch of this job before calling the HUDOperator.
-    // If not, the current Framebuffer is left unchanged.
-    //using Inputs = gpu::FramebufferPointer;
-    using JobModel = render::Job::ModelI<CompositeHUD, gpu::FramebufferPointer>;
-
-    void run(const render::RenderContextPointer& renderContext, const gpu::FramebufferPointer& inputs);
-};
-
 class Blit {
 public:
     using JobModel = render::Job::ModelI<Blit, gpu::FramebufferPointer>;
